OTC-624 refactoring and cleaning up test setup
