+++ package.json
@@ -35,7 +35,7 @@
         "@mojaloop/central-services-error-handling": "10.4.1",
         "@mojaloop/central-services-health": "10.4.0",
         "@mojaloop/central-services-logger": "10.4.0",
-        "@mojaloop/central-services-shared": "10.5.2",
+        "@mojaloop/central-services-shared": "10.5.3",
         "@mojaloop/central-services-stream": "9.5.0",
         "@mojaloop/ml-number": "8.2.0",
         "@now-ims/hapi-now-auth": "2.0.2",
@@ -54,11 +54,11 @@
     },
     "devDependencies": {
         "@hapi/joi": "17.1.1",
+        "axios": "^0.19.2",
         "bluebird": "3.7.2",
-        "eslint": "7.3.1",
+        "eslint": "7.4.0",
         "faucet": "0.0.1",
         "get-port": "5.1.1",
-        "node-fetch": "2.6.0",
         "nodemon": "2.0.4",
         "npm-audit-resolver": "2.2.0",
         "npm-check-updates": "6.0.1",
